Remove old async message pump that tries to use STA approach; to be replaced with various message pump approaches other than STA which include async options
