+++ src/Paramore.Brighter.ServiceActivator/IAmAMessagePump.cs
@@ -51,10 +51,5 @@ namespace Paramore.Brighter.ServiceActivator
         /// </summary>
         /// <value>The channel.</value>
         IAmAChannel Channel { get; set; }
-
-        /// <summary>
-        /// Is this message pump for an async pipeline, or a synchronous pipeline
-        /// </summary>
-        bool IsAsync { get; }
-    }
+   }
 }
