+++ tests/Paramore.Brighter.Core.Tests/MessageDispatch/When_a_message_dispatcher_Is_asked_to_connect_a_channel_and_handler_async.cs
@@ -40,21 +40,5 @@ namespace Paramore.Brighter.Core.Tests.MessageDispatch
             _dispatcher.State.Should().Be(DispatcherState.DS_AWAITING);
             _dispatcher.Receive();
         }
-
-        [Fact]
-        public void When_a_message_dispatcher_is_asked_to_connect_a_channel_and_handler_async()
-        {
-            Task.Delay(1000).Wait();
-            _dispatcher.End().Wait();
-
-            //_should_have_consumed_the_messages_in_the_channel
-            _channel.Length.Should().Be(0);
-            //_should_have_a_stopped_state
-            _dispatcher.State.Should().Be(DispatcherState.DS_STOPPED);
-            //_should_have_dispatched_a_request
-            _commandProcessor.Observe<MyEvent>().Should().NotBeNull();
-            //_should_have_published_async
-            _commandProcessor.Commands.Should().Contain(ctype => ctype == CommandType.PublishAsync);
-        }
-    }
+   }
 }
